+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - added Body.move_to_origin() Body.center() and Body.mirror()
 - added Sketch.move_to_origin() Sketch.center() and Sketch.mirror()
 - added pattern_distribute()
+- breaking change: Sketch.add_polygon() has now parameter auto_close which is True by default
 
 ## [0.4] - 2024-01-12